requirements.txt
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
fonttools==4.58.0
fsspec==2024.6.1
Jinja2==3.1.4
joblib==1.5.0
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==25.0
pillow==11.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
scipy==1.15.3
setuptools==70.2.0
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.7.0+cpu
torchaudio==2.7.0+cpu
torchvision==0.22.0+cpu
tqdm==4.67.1
typing_extensions==4.12.2
wheel==0.45.1
